Move Conditions to CorgiLib
